pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "dataclasses-json>=0.6.7",
    "firecrawl-py>=1.12.0",
    "google-api-python-client>=2.161.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "langchain-anthropic>=0.3.8",
    "langchain-community>=0.3.18",
    "langchain-core>=0.3.40",
    "langchain-google-vertexai>=2.0.14",
    "langchain-openai>=0.3.7",
    "numpy>=2.2.3",
    "openai>=1.64.0",
    "pandas>=2.2.3",
    "plotly[express]>=6.0.0",
    "pydantic>=2.10.6",
    "pymongo>=4.11.1",
    "pystache>=0.6.7",
    "pytest-playwright>=0.7.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "scipy>=1.15.2",
    "twelvedata[pandas]>=1.2.24",
    "xlsxwriter>=3.2.2",
]

[tool.ruff]
line-length = 256

[tool.ruff.lint]
extend-select = ["E", "F", "N", "R", "I", "UP"]
